+++ example/example.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:an_lifecycle_cancellable/an_lifecycle_cancellable.dart';
 import 'package:an_lifecycle_viewmodel/an_lifecycle_viewmodel.dart';
 import 'package:anlifecycle/anlifecycle.dart';
@@ -17,14 +15,14 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return LifecycleApp(
       child: MaterialApp(
-        title: 'Lifecycle ViewModel Demo',
+        title: 'ViewModel Demo',
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
         navigatorObservers: [
           LifecycleNavigatorObserver.hookMode(),
         ],
-        home: const MyHomePage(title: 'Lifecycle ViewModel Demo Home Page'),
+        home: const MyHomePage(title: 'ViewModel Demo Home Page'),
       ),
     );
   }
@@ -51,32 +49,14 @@ class MyHomePage extends StatefulWidget {
   State<MyHomePage> createState() => _MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage>
-    with LifecycleRegistryStateMixin {
-  late final ValueNotifier<int> stayed = ValueNotifier(0);
-
-  late final ViewModelHome viewModel = viewModels();
+class _MyHomePageState extends State<MyHomePage> {
+  // 获取当前环境下的ViewModel
+  late final ViewModelHome viewModel = viewModelsOfState();
 
   // 也可使用 当前注册的构建工厂
   // final viewModel =
   //     useLifecycleViewModelEffect<ViewModelHome>(factory2: ViewModelHome.new);
   // late final ViewModelHome viewModel = viewModels(factory2: ViewModelHome.new);
-
-  @override
-  void initState() {
-    super.initState();
-    stayed
-        .bindLifecycle(this)
-        .addCListener(makeLiveCancellable(), () => setState(() {}));
-
-    // 只有可见时统计时间 不可见时不统计
-    Stream.periodic(const Duration(seconds: 1))
-        .bindLifecycle(this, repeatLastOnRestart: true)
-        .listen((event) => stayed.value++);
-
-    viewModel.counter
-        .addCListener(makeLiveCancellable(), () => setState(() {}));
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -88,15 +68,15 @@ class _MyHomePageState extends State<MyHomePage>
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            Text(
-              'Stayed on this page for:${stayed.value} s',
-            ),
             const Text(
               'You have pushed the button this many times:',
             ),
-            Text(
-              '${viewModel.counter.value}',
-              style: Theme.of(context).textTheme.headlineMedium,
+            AnimatedBuilder(
+              animation: viewModel.counter,
+              builder: (_, __) => Text(
+                '${viewModel.counter.value}',
+                style: Theme.of(context).textTheme.headlineMedium,
+              ),
             ),
           ],
         ),
@@ -114,12 +94,12 @@ class HomeFloatingButton extends StatefulWidget {
   State<HomeFloatingButton> createState() => _HomeFloatingButtonState();
 }
 
-class _HomeFloatingButtonState extends State<HomeFloatingButton>
-    with LifecycleRegistryStateMixin {
+class _HomeFloatingButtonState extends State<HomeFloatingButton> {
+  //获取vm
+  late final vm = viewModelsOfState<ViewModelHome>();
+
   @override
   Widget build(BuildContext context) {
-    //获取vm
-    final vm = viewModels<ViewModelHome>();
     return FloatingActionButton(
       onPressed: vm.incrementCounter,
       tooltip: 'Increment',